Align native preview CLI package layout
