+++ .dprint.jsonc
@@ -52,7 +52,7 @@
         "internal/lsp/lsproto/_generate/*.json",
         "internal/lsp/lsproto/_generate/metaModelSchema.mts",
         // Needs to be LF to have a working shebang.
-        "_packages/native-preview/bin/tsgo.js",
+        "_packages/native-preview/bin/tsgo",
         "_packages/native-preview/vendor/**"
     ],
     // Note: if adding new languages, make sure settings.template.json is updated too.
