+++ _packages/native-preview/bin/tsgo
@@ -1,0 +1,2 @@
+#!/usr/bin/env node
+import "../lib/tsgo.js";
