+++ _packages/native-preview/package.json
@@ -36,7 +36,7 @@
         "vendor"
     ],
     "bin": {
-        "tsgo": "./bin/tsgo.js"
+        "tsgo": "./bin/tsgo"
     },
     "exports": {
         "./package.json": "./package.json",
